requirements.txt
sanic==25.3.0
pydantic==2.11.3
aiohttp==3.11.16
python-dotenv==1.1.0
psycopg-binary==3.2.6
psycopg==3.2.6
